Move the asynhronous tests to .Async partial classes
