+++ test/FunctionalTests/Microsoft.OData.Edm.Tests/Csdl/Serialization/EdmModelCsdlSchemaWriterTests.cs
@@ -6,7 +6,6 @@
 
 using System;
 using System.IO;
-using System.Threading.Tasks;
 using System.Xml;
 using Microsoft.OData.Edm.Csdl;
 using Microsoft.OData.Edm.Csdl.Serialization;
@@ -18,7 +17,7 @@ namespace Microsoft.OData.Edm.Tests.Csdl.Serialization
     /// <summary>
     ///Tests EdmModelCsdlSchemaWriter functionalities
     ///</summary>
-    public class EdmModelCsdlSchemaWriterTests
+    public partial class EdmModelCsdlSchemaWriterTests
     {
         private static readonly EdmEntityContainer defaultContainer = new EdmEntityContainer("Default.NameSpace", "Container");
         private static readonly EdmAction defaultCheckoutAction = new EdmAction("Default.NameSpace2", "CheckOut", null);
@@ -34,27 +33,11 @@ namespace Microsoft.OData.Edm.Tests.Csdl.Serialization
         }
 
         [Fact]
-        public async Task BoundOperationShouldWriteIsBoundEqualTrueAttribute_Async()
-        {
-            EdmAction action = new EdmAction("Default.Namespace", "Checkout", null /*returnType*/, true /*isBound*/, null /*entitySetPath*/);
-            action.AddParameter("param", EdmCoreModel.Instance.GetString(true));
-            await this.TestWriteActionElementHeaderMethodWithAsync(action, @"<Action Name=""Checkout"" IsBound=""true""").ConfigureAwait(false);
-        }
-
-        [Fact]
         public void NonBoundOperationShouldNotWriteIsBoundAttribute()
         {
             EdmFunction function = new EdmFunction("Default.Namespace", "Checkout", EdmCoreModel.Instance.GetString(true) /*returnType*/, false /*isBound*/, null /*entitySetPath*/, false /*isComposable*/);
             function.AddParameter("param", EdmCoreModel.Instance.GetString(true));
             this.TestWriteFunctionElementHeaderMethod(function, @"<Function Name=""Checkout""");
-        }
-
-        [Fact]
-        public async Task NonBoundOperationShouldNotWriteIsBoundAttribute_Async()
-        {
-            EdmFunction function = new EdmFunction("Default.Namespace", "Checkout", EdmCoreModel.Instance.GetString(true) /*returnType*/, false /*isBound*/, null /*entitySetPath*/, false /*isComposable*/);
-            function.AddParameter("param", EdmCoreModel.Instance.GetString(true));
-            await this.TestWriteFunctionElementHeaderMethodWithAsync(function, @"<Function Name=""Checkout""").ConfigureAwait(false);
         }
 
         [Fact]
@@ -66,25 +49,10 @@ namespace Microsoft.OData.Edm.Tests.Csdl.Serialization
         }
 
         [Fact]
-        public async Task BoundOperationWithEntitySetPathShouldWriteEntitySetPathAttributeWithCorrectValue_Async()
-        {
-            EdmAction action = new EdmAction("Default.Namespace", "Checkout", null /*returnType*/, true /*isBound*/, new EdmPathExpression("Customer", "Orders") /*entitySetPath*/);
-            action.AddParameter("param", EdmCoreModel.Instance.GetString(true));
-            await this.TestWriteActionElementHeaderMethodWithAsync(action, @"<Action Name=""Checkout"" IsBound=""true"" EntitySetPath=""Customer/Orders""").ConfigureAwait(false);
-        }
-
-        [Fact]
         public void ComposableFunctionShouldWriteIsComposableEqualTrue()
         {
             EdmFunction function = new EdmFunction("Default.Namespace", "Checkout", EdmCoreModel.Instance.GetString(true) /*returnType*/, false /*isBound*/, null /*entitySetPath*/, true /*isComposable*/);
             this.TestWriteFunctionElementHeaderMethod(function, @"<Function Name=""Checkout"" IsComposable=""true""");
-        }
-
-        [Fact]
-        public async Task ComposableFunctionShouldWriteIsComposableEqualTrue_Async()
-        {
-            EdmFunction function = new EdmFunction("Default.Namespace", "Checkout", EdmCoreModel.Instance.GetString(true) /*returnType*/, false /*isBound*/, null /*entitySetPath*/, true /*isComposable*/);
-            await this.TestWriteFunctionElementHeaderMethodWithAsync(function, @"<Function Name=""Checkout"" IsComposable=""true""").ConfigureAwait(false);
         }
 
         private void TestWriteActionElementHeaderMethod(EdmAction action, string expected)
@@ -92,19 +60,9 @@ namespace Microsoft.OData.Edm.Tests.Csdl.Serialization
             this.EdmModelCsdlSchemaWriterTest(writer => writer.WriteActionElementHeader(action), expected);
         }
 
-        private async Task TestWriteActionElementHeaderMethodWithAsync(EdmAction action, string expected)
-        {
-            await this.EdmModelCsdlSchemaWriterTestAsync(async (writer) => await writer.WriteActionElementHeaderAsync(action).ConfigureAwait(false), expected).ConfigureAwait(false);
-        }
-
         private void TestWriteFunctionElementHeaderMethod(EdmFunction function, string expected)
         {
             this.EdmModelCsdlSchemaWriterTest(writer => writer.WriteFunctionElementHeader(function), expected);
-        }
-
-        private async Task TestWriteFunctionElementHeaderMethodWithAsync(EdmFunction function, string expected)
-        {
-            await this.EdmModelCsdlSchemaWriterTestAsync(async (writer) => await writer.WriteFunctionElementHeaderAsync(function).ConfigureAwait(false), expected).ConfigureAwait(false);
         }
         #endregion
 
@@ -119,32 +77,12 @@ namespace Microsoft.OData.Edm.Tests.Csdl.Serialization
         }
 
         [Fact]
-        public async Task ValidateEntitySetAtttributeCorrectlyWritesOutEntitySet_Async()
-        {
-            EdmActionImport actionImport = new EdmActionImport(defaultContainer, "Checkout", defaultCheckoutAction, new EdmPathExpression("Customers"));
-
-            await this.EdmModelCsdlSchemaWriterTestAsync(
-                async (writer) => await writer.WriteActionImportElementHeaderAsync(actionImport).ConfigureAwait(false),
-                @"<ActionImport Name=""Checkout"" Action=""Default.NameSpace2.CheckOut"" EntitySet=""Customers""").ConfigureAwait(false);
-        }
-
-        [Fact]
         public void ValidateEntitySetAtttributeCorrectlyWritesEntitySetPath()
         {
             EdmFunctionImport functionImport = new EdmFunctionImport(defaultContainer, "GetStuff", defaultGetStuffFunction, new EdmPathExpression("Customers", "Orders"), false);
 
             this.EdmModelCsdlSchemaWriterTest(writer => writer.WriteFunctionImportElementHeader(functionImport),
                 @"<FunctionImport Name=""GetStuff"" Function=""Default.NameSpace2.GetStuff"" EntitySet=""Customers/Orders""");
-        }
-
-        [Fact]
-        public async Task ValidateEntitySetAtttributeCorrectlyWritesEntitySetPath_Async()
-        {
-            EdmFunctionImport functionImport = new EdmFunctionImport(defaultContainer, "GetStuff", defaultGetStuffFunction, new EdmPathExpression("Customers", "Orders"), false);
-
-            await this.EdmModelCsdlSchemaWriterTestAsync(
-                async (writer) => await writer.WriteFunctionImportElementHeaderAsync(functionImport).ConfigureAwait(false),
-                @"<FunctionImport Name=""GetStuff"" Function=""Default.NameSpace2.GetStuff"" EntitySet=""Customers/Orders""").ConfigureAwait(false);
         }
 
         [Fact]
@@ -155,18 +93,6 @@ namespace Microsoft.OData.Edm.Tests.Csdl.Serialization
             Action errorTest = () => CreateEdmModelCsdlSchemaWriterForErrorTest().WriteActionImportElementHeader(actionImport);
 
             var exception = Assert.Throws<InvalidOperationException>(errorTest);
-            Assert.Equal(Strings.EdmModel_Validator_Semantic_OperationImportEntitySetExpressionIsInvalid(actionImport.Name), exception.Message);
-        }
-
-        [Fact]
-        public async Task ValidateIncorrectEdmExpressionThrows_Async()
-        {
-            EdmActionImport actionImport = new EdmActionImport(defaultContainer, "Checkout", defaultCheckoutAction, new EdmIntegerConstant(EdmCoreModel.Instance.GetInt32(true), 1));
-            
-            var csdlSchemaWriter = CreateEdmModelCsdlSchemaWriterForErrorTest();
-            async Task errorTest() => await csdlSchemaWriter.WriteActionImportElementHeaderAsync(actionImport).ConfigureAwait(false);
-
-            var exception = await Assert.ThrowsAsync<InvalidOperationException>(errorTest).ConfigureAwait(false);
             Assert.Equal(Strings.EdmModel_Validator_Semantic_OperationImportEntitySetExpressionIsInvalid(actionImport.Name), exception.Message);
         }
 
@@ -182,16 +108,6 @@ namespace Microsoft.OData.Edm.Tests.Csdl.Serialization
                 @"<ActionImport Name=""Checkout"" Action=""Default.NameSpace2.CheckOut""");
         }
 
-        [Fact]
-        public async Task ValidateCorrectActionImportNameAndActionAttributeValueWrittenCorrectly_Async()
-        {
-            EdmActionImport actionImport = new EdmActionImport(defaultContainer, "Checkout", defaultCheckoutAction, null);
-
-            await this.EdmModelCsdlSchemaWriterTestAsync(
-                async (writer) => await writer.WriteActionImportElementHeaderAsync(actionImport).ConfigureAwait(false),
-                @"<ActionImport Name=""Checkout"" Action=""Default.NameSpace2.CheckOut""").ConfigureAwait(false);
-        }
-
         #endregion
 
         #region FunctionImport tests.
@@ -205,32 +121,12 @@ namespace Microsoft.OData.Edm.Tests.Csdl.Serialization
         }
 
         [Fact]
-        public async Task ValidateCorrectFunctionNameAndFunctionAttributeValueWrittenCorrectly_Async()
-        {
-            EdmFunctionImport functionImport = new EdmFunctionImport(defaultContainer, "GetStuff", defaultGetStuffFunction, new EdmPathExpression("Customers", "Orders"), false);
-
-            await this.EdmModelCsdlSchemaWriterTestAsync(
-                async (writer) => await writer.WriteFunctionImportElementHeaderAsync(functionImport).ConfigureAwait(false),
-                @"<FunctionImport Name=""GetStuff"" Function=""Default.NameSpace2.GetStuff"" EntitySet=""Customers/Orders""").ConfigureAwait(false);
-        }
-
-        [Fact]
         public void ValidateFunctionIncludeInServiceDocumentWrittenAsTrue()
         {
             EdmFunctionImport functionImport = new EdmFunctionImport(defaultContainer, "GetStuff", defaultGetStuffFunction, new EdmPathExpression("Customers", "Orders"), true);
 
             this.EdmModelCsdlSchemaWriterTest(writer => writer.WriteFunctionImportElementHeader(functionImport),
                 @"<FunctionImport Name=""GetStuff"" Function=""Default.NameSpace2.GetStuff"" EntitySet=""Customers/Orders"" IncludeInServiceDocument=""true""");
-        }
-
-        [Fact]
-        public async Task ValidateFunctionIncludeInServiceDocumentWrittenAsTrue_Async()
-        {
-            EdmFunctionImport functionImport = new EdmFunctionImport(defaultContainer, "GetStuff", defaultGetStuffFunction, new EdmPathExpression("Customers", "Orders"), true);
-
-            await this.EdmModelCsdlSchemaWriterTestAsync(
-                async (writer) => await writer.WriteFunctionImportElementHeaderAsync(functionImport).ConfigureAwait(false),
-                @"<FunctionImport Name=""GetStuff"" Function=""Default.NameSpace2.GetStuff"" EntitySet=""Customers/Orders"" IncludeInServiceDocument=""true""").ConfigureAwait(false);
         }
         #endregion
 
@@ -242,21 +138,9 @@ namespace Microsoft.OData.Edm.Tests.Csdl.Serialization
             TestWriteComplexTypeElementHeaderMethod(complexType, @"<ComplexType Name=""OpenComplex"" OpenType=""true""");
         }
 
-        [Fact]
-        public async Task ShouldWriteOpenTypeAttributeForOpenComplexType_Async()
-        {
-            IEdmComplexType complexType = new EdmComplexType("Default.NameSpace2", "OpenComplex", null, false, true);
-            await TestWriteComplexTypeElementHeaderMethodWithAsync(complexType, @"<ComplexType Name=""OpenComplex"" OpenType=""true""").ConfigureAwait(false);
-        }
-
         private void TestWriteComplexTypeElementHeaderMethod(IEdmComplexType complexType, string expected)
         {
             this.EdmModelCsdlSchemaWriterTest(writer => writer.WriteComplexTypeElementHeader(complexType), expected);
-        }
-
-        private async Task TestWriteComplexTypeElementHeaderMethodWithAsync(IEdmComplexType complexType, string expected)
-        {
-            await this.EdmModelCsdlSchemaWriterTestAsync(async (writer) => await writer.WriteComplexTypeElementHeaderAsync(complexType).ConfigureAwait(false), expected).ConfigureAwait(false);
         }
         #endregion
 
@@ -269,24 +153,6 @@ namespace Microsoft.OData.Edm.Tests.Csdl.Serialization
             testAction(schemaWriter);
 
             xmlWriter.Flush();
-            memoryStream.Seek(0, SeekOrigin.Begin);
-            StreamReader reader = new StreamReader(memoryStream);
-
-            // Removing xml header to make the baseline's more compact and focused on the test at hand.
-            string result = reader.ReadToEnd().Replace(@"<?xml version=""1.0"" encoding=""utf-8""?>", string.Empty);
-            Assert.Equal(expectedPayload, result);
-        }
-
-        internal async Task EdmModelCsdlSchemaWriterTestAsync(Func<EdmModelCsdlSchemaWriter, Task> testAction, string expectedPayload)
-        {
-            XmlWriter xmlWriter;
-            MemoryStream memoryStream;
-            EdmModelCsdlSchemaWriter schemaWriter = CreateEdmModelCsdlSchemaWriter(out xmlWriter, out memoryStream);
-
-            await testAction(schemaWriter).ConfigureAwait(false);
-
-            await xmlWriter.FlushAsync().ConfigureAwait(false);
-
             memoryStream.Seek(0, SeekOrigin.Begin);
             StreamReader reader = new StreamReader(memoryStream);
