+++ test/FunctionalTests/Microsoft.OData.Edm.Tests/Vocabularies/CommunityVocabularyTests.Async.cs
@@ -11,18 +11,16 @@ using System.Xml;
 using System.Text;
 using Microsoft.OData.Edm.Csdl;
 using Microsoft.OData.Edm.Validation;
-using Microsoft.OData.Edm.Vocabularies;
-using Microsoft.OData.Edm.Vocabularies.Community.V1;
 using Xunit;
+using System.Threading.Tasks;
 
 namespace Microsoft.OData.Edm.Tests.Vocabularies
 {
     public partial class CommunityVocabularyTests
     {
-        private readonly IEdmModel model = CommunityVocabularyModel.Instance;
 
         [Fact]
-        public void TestCommunityVocabularyModel()
+        public async Task TestCommunityVocabularyModel_Async()
         {
             const string expectedUrlEscape = @"<?xml version=""1.0"" encoding=""utf-16""?>
 <Schema Namespace=""Org.OData.Community.V1"" Alias=""Community"" xmlns=""http://docs.oasis-open.org/odata/ns/edm"">
@@ -31,28 +29,20 @@ namespace Microsoft.OData.Edm.Tests.Vocabularies
   </Term>
 </Schema>";
 
-            StringWriter sw = new StringWriter();
+            var sw = new StringWriter();
             IEnumerable<EdmError> errors;
-            using (var xw = XmlWriter.Create(sw, new XmlWriterSettings { Indent = true, Encoding = Encoding.UTF8 }))
+            using (var xw = XmlWriter.Create(sw, new XmlWriterSettings { Indent = true, Encoding = Encoding.UTF8, Async = true }))
             {
-                Assert.True(model.TryWriteSchema(xw, out errors));
-            }
+                var (result, errorsAsync) = await model.TryWriteSchemaAsync(xw).ConfigureAwait(false);
+                Assert.True(result);
 
+                errors = errorsAsync;
+            }
 
             Assert.False(errors.Any(), "No Errors");
 
             string output = sw.ToString();
             Assert.True(expectedUrlEscape == output, "expected Community schema not matching");
-        }
-
-        [Fact]
-        public void TestUrlEscapeFunctionAnnotationTerm()
-        {
-            var urlEscapeFunctionTerm = model.FindDeclaredTerm(CommunityVocabularyConstants.UrlEscapeFunction);
-            Assert.NotNull(urlEscapeFunctionTerm);
-            Assert.Equal(CommunityVocabularyModel.UrlEscapeFunctionTerm, urlEscapeFunctionTerm);
-            Assert.Equal("Org.OData.Community.V1", urlEscapeFunctionTerm.Namespace);
-            Assert.Equal("UrlEscapeFunction", urlEscapeFunctionTerm.Name);
         }
     }
 }
